# schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Local Community Back End Demonstration API
  description: A Local Community API, used to register user, create new issue, organise meetings and deliver announcements
  contact:
    name: Harman Singh
    email: singhh48@coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://outsidefilter-epoxyderby-8080.codio-box.uk/api/v1'
    description: Development server
...
paths:
...
  /issues:
    summary: Rappresents meetings posts on the database
    description: >
      This resource represents an issue post in the system.
      Each post is identified by a numeric `id`.
    get:
      tags:
        - meetings
      description:
        Get all issues.
      responses:
        '200':
          description: View all issues record
          content:
            application/json:
              schema:
                $ref: ./issue.json#/definitions/issues
        '401':
          description: Non-admin users can all data
        '404':
          description: Not found

    post:
      tags:
        - Issues
      description:
        Create a Issue Post.
      requestBody:
        description: All issues post body data.
        required: true
        content:
          application/json:
            schema:
              $ref: ./issues.json#/definitions/issues
      responses:
        '201':
          description: Issue post created
          content:
            application/json:
              schema:
                $ref: ./issues.json#/definitions/issues
  /issues/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the issues post
    get:
      tags:
        - Issues
      description:
        Get issues by id.
      responses:
        '200':
          description: View of issue record
          content:
            application/json:
              schema:
                $ref: ./issues.json#/definitions/issues
        '401':
          description: Non-admin users can view their any data
        '404':
          description: Not foundd

    put:
      tags:
        - Issues
      description:
        Update issue post by id.
      requestBody:
        description: Update a visible property.
        required: true
        content:
          application/json:
            schema:
              $ref: ./issues.json#/definitions/issueUpdate
      responses:
        '200':
          description: Confirmation and link to record.
          content:
            application/json:
              schema:
                $ref: ./issue.json#/definitions/issueUpdated
        '404':
          description: Not found
...